Refactor the implementation
